+++ pkgs/sse/lib/src/server/sse_handler.dart
@@ -30,6 +30,8 @@ class _SseMessage {
 
 /// A bi-directional SSE connection between server and browser.
 class SseConnection extends StreamChannelMixin<String> {
+  final _logger = Logger('SseConnection');
+
   /// Incoming messages from the Browser client.
   final _incomingController = StreamController<String>();
 
@@ -57,9 +59,16 @@ class SseConnection extends StreamChannelMixin<String> {
 
   final _closedCompleter = Completer<void>();
 
+  /// Whether the connection is permanently closed.
+  bool get isClosed => _closedCompleter.isCompleted;
+
   /// Wraps the `_outgoingController.stream` to buffer events to enable keep
   /// alive.
   late StreamQueue _outgoingStreamQueue;
+
+  final Duration? _ignoreDisconnect;
+
+  final _connectionStopwatch = Stopwatch();
 
   /// Creates an [SseConnection] for the supplied [_sink].
   ///
@@ -70,7 +79,13 @@ class SseConnection extends StreamChannelMixin<String> {
   ///
   /// If [keepAlive] is not supplied, the connection will be closed immediately
   /// after a disconnect.
-  SseConnection(this._sink, {Duration? keepAlive}) : _keepAlive = keepAlive {
+  ///
+  /// [ignoreDisconnect] is the duration to ignore a disconnect after a
+  /// reconnection. This is useful as the SSE disconnection event may arrive
+  /// after the new connection is established.
+  SseConnection(this._sink, {Duration? keepAlive, Duration? ignoreDisconnect})
+    : _keepAlive = keepAlive,
+      _ignoreDisconnect = ignoreDisconnect {
     _outgoingStreamQueue = StreamQueue(_outgoingController.stream);
     unawaited(_setUpListener());
     _outgoingController.onCancel = _close;
@@ -140,21 +155,54 @@ class SseConnection extends StreamChannelMixin<String> {
       } else {
         // A message came out of order. Wait until we receive the previous
         // messages to process.
+        _logger.warning(
+          'Message received out of order, last processed id: $_lastProcessedId,'
+          ' current message id: ${pendingMessage.id}',
+        );
         break;
       }
     }
   }
 
   void _acceptReconnection(Sink sink) {
+    if (!isInKeepAlivePeriod) {
+      _logger.warning(
+        'Accepting reconnection while not in keep alive period, this usually '
+        'happens when the new connection request arrives before the SSE '
+        'disconnection event.',
+      );
+
+      if (_ignoreDisconnect != null) {
+        _connectionStopwatch
+          ..reset()
+          ..start();
+      }
+    }
+
     _keepAliveTimer?.cancel();
     _sink = sink;
   }
 
   void _handleDisconnect() {
+    // Determine how long ago we were last reconnected.
+    // Ignore disconnect if we just recently reconnected. This happens because
+    // the disconnect event arrives after the connect event.
+    if (_ignoreDisconnect != null) {
+      _connectionStopwatch.stop();
+      if (_ignoreDisconnect > _connectionStopwatch.elapsed) {
+        _logger.warning(
+          'Ignoring disconnect ${_connectionStopwatch.elapsed.inMilliseconds} '
+          'milliseconds after connection which is less than the '
+          '${_ignoreDisconnect.inMilliseconds} millisecond minimum.',
+        );
+        return;
+      }
+    }
+
     if (_keepAlive == null) {
       // Close immediately if we're not keeping alive.
       _close();
-    } else if (!isInKeepAlivePeriod && !_closedCompleter.isCompleted) {
+    } else if (!isInKeepAlivePeriod && !isClosed) {
       // Otherwise if we didn't already have an active timer and we've not
       // already been completely closed, set a timer to close after the timeout
       // period.
@@ -196,6 +244,7 @@ class SseHandler {
   final Duration? _keepAlive;
   final _connections = <String?, SseConnection>{};
   final _connectionController = StreamController<SseConnection>();
+  final Duration? _ignoreDisconnect;
 
   StreamQueue<SseConnection>? _connectionsStream;
 
@@ -209,7 +258,13 @@ class SseHandler {
   ///
   /// If [keepAlive] is not supplied, connections will be closed immediately
   /// after a disconnect.
-  SseHandler(this._uri, {Duration? keepAlive}) : _keepAlive = keepAlive;
+  ///
+  /// If [ignoreDisconnect] is supplied, it will be passed on to the
+  /// [SseConnection]. The SseHandler will also use it as a signal to reuse
+  /// open SSE connections that are not in a keep alive period.
+  SseHandler(this._uri, {Duration? keepAlive, Duration? ignoreDisconnect})
+    : _keepAlive = keepAlive,
+      _ignoreDisconnect = ignoreDisconnect;
 
   StreamQueue<SseConnection> get connections =>
       _connectionsStream ??= StreamQueue(_connectionController.stream);
@@ -224,14 +279,22 @@ class SseHandler {
       sink.add(_sseHeaders(req.headers['origin']));
       var clientId = req.url.queryParameters['sseClientId'];
 
-      // Check if we already have a connection for this ID that is in the
-      // process of timing out
+      // Check if we already have a connection for this ID that is either:
+      // - in the process of timing out
+      // - not closed and we're ignoring disconnects.
+      //
       // (in which case we can reconnect it transparently).
       if (_connections[clientId] != null &&
-          _connections[clientId]!.isInKeepAlivePeriod) {
+          (_connections[clientId]!.isInKeepAlivePeriod ||
+              (_ignoreDisconnect != null &&
+                  !_connections[clientId]!.isClosed))) {
         _connections[clientId]!._acceptReconnection(sink);
       } else {
-        var connection = SseConnection(sink, keepAlive: _keepAlive);
+        var connection = SseConnection(
+          sink,
+          keepAlive: _keepAlive,
+          ignoreDisconnect: _ignoreDisconnect,
+        );
         _connections[clientId] = connection;
         unawaited(
           connection._closedCompleter.future.then((_) {