+++ pkgs/sse/CHANGELOG.md
@@ -1,5 +1,10 @@
-## 4.2.0
+## 4.2.0-wip
 
+- Handle out of order disconnects and reconnects.
+  - Allow SSE connections to be reused if not in a keep alive period.
+  - Allow a debounce period after connections to ignore disconnects which may be
+    from prior connections.
+  - Add logging when messages are received out of order.
 - Update the minimum Dart SDK version to `3.11.0`.
 
 ## 4.1.8